+++ forge/src/main/java/de/maxhenkel/voicechat/ForgeVoicechatMod.java
@@ -1,35 +1,30 @@
 package de.maxhenkel.voicechat;
 
-import de.maxhenkel.voicechat.config.ForgeServerConfig;
+import de.maxhenkel.voicechat.config.ConfigMigrator;
 import de.maxhenkel.voicechat.intercompatibility.CommonCompatibilityManager;
 import de.maxhenkel.voicechat.permission.PermissionManager;
 import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.DistExecutor;
 import net.minecraftforge.fml.IExtensionPoint;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
 import java.util.Objects;
-import java.util.function.Function;
 
 @Mod(ForgeVoicechatMod.MODID)
 public class ForgeVoicechatMod extends Voicechat {
 
     public ForgeVoicechatMod() {
         FMLJavaModLoadingContext.get().getModEventBus().addListener(this::commonSetup);
-
-        SERVER_CONFIG = registerConfig(ModConfig.Type.SERVER, ForgeServerConfig::new);
-
         DistExecutor.unsafeRunWhenOn(Dist.CLIENT, () -> ForgeVoicechatClientMod::new);
     }
 
     public void commonSetup(FMLCommonSetupEvent event) {
         initialize();
+        MinecraftForge.EVENT_BUS.register(new ConfigMigrator());
         MinecraftForge.EVENT_BUS.register(CommonCompatibilityManager.INSTANCE);
         MinecraftForge.EVENT_BUS.register(PermissionManager.INSTANCE);
         ModLoadingContext.get().registerExtensionPoint(IExtensionPoint.DisplayTest.class, () -> {
@@ -39,11 +34,4 @@ public class ForgeVoicechatMod extends Voicechat {
         });
     }
 
-    public static <T> T registerConfig(ModConfig.Type type, Function<ForgeConfigSpec.Builder, T> consumer) {
-        ForgeConfigSpec.Builder builder = new ForgeConfigSpec.Builder();
-        T config = consumer.apply(builder);
-        ForgeConfigSpec spec = builder.build();
-        ModLoadingContext.get().registerConfig(type, spec);
-        return config;
-    }
 }