+++ quilt/src/main/java/de/maxhenkel/voicechat/QuiltVoicechatClientMod.java
@@ -1,10 +1,5 @@
 package de.maxhenkel.voicechat;
 
-import com.sun.jna.Platform;
-import de.maxhenkel.configbuilder.ConfigBuilder;
-import de.maxhenkel.voicechat.config.QuiltClientConfig;
-import de.maxhenkel.voicechat.macos.VersionCheck;
-import net.minecraft.client.Minecraft;
 import org.quiltmc.loader.api.ModContainer;
 import org.quiltmc.qsl.base.api.entrypoint.client.ClientModInitializer;
 
@@ -12,14 +7,6 @@ public class QuiltVoicechatClientMod extends VoicechatClient implements ClientMo
 
     @Override
     public void onInitializeClient(ModContainer mod) {
-        CLIENT_CONFIG = ConfigBuilder.build(Minecraft.getInstance().gameDirectory.toPath().resolve("config").resolve(Voicechat.MODID).resolve("voicechat-client.properties"), true, QuiltClientConfig::new);
         initializeClient();
-
-        if (Platform.isMac() && !CLIENT_CONFIG.javaMicrophoneImplementation.get()) {
-            CLIENT_CONFIG.javaMicrophoneImplementation.set(true).save();
-        }
-        if (Platform.isMac() && CLIENT_CONFIG.useNatives.get() && !VersionCheck.isMacOSNativeCompatible()) {
-            CLIENT_CONFIG.useNatives.set(false).save();
-        }
     }
 }